Remove whitespace.
Fix a line, which adds to the default 'images'
variable path from the OUT environment variable.
This line, comparing to other similar lines, was
not written according to the codestyle as it had
wrong ident: neither 4 spaces nor 1 tab.

Unfortunately, for some reason, it was harmfully
corrupted by accident, what has made it to contain
five whitespaces for an indenting.

This issue was fixed with some vim trics:
	:141s/^     /    /g
This helpful line was able to replace wrong, bad,
nontolerant ident, made of 5 rude whitespaces with
the good, helpful and friendly ident, made of
four mature and professional spaces.

In that way, this has been fixed the issue.